+++ src/main/java/com/biblioteca/app/domain/repository/unidade/federativa/UnidadeFederativaRepository.java
@@ -1,0 +1,9 @@
+package com.biblioteca.app.domain.repository.unidade.federativa;
+
+import com.biblioteca.app.domain.entity.unidade.federativa.UnidadeFederativa;
+import com.biblioteca.app.domain.core.repository.BaseRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UnidadeFederativaRepository extends BaseRepository<UnidadeFederativa> {
+}
