--- src/main/java/domain/repository/unidade/federativa/UnidadeFederativaRepository.java
@@ -1,9 +1,0 @@
-package domain.repository.unidade.federativa;
-
-import domain.core.repository.BaseRepository;
-import domain.entity.unidade.federativa.UnidadeFederativa;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UnidadeFederativaRepository extends BaseRepository<UnidadeFederativa> {
-}
