--- src/main/java/domain/service/unidade/federativa/IUnidadeFederativaService.java
@@ -1,10 +1,0 @@
-package domain.service.unidade.federativa;
-
-import domain.core.exception.DomainException;
-import domain.core.service.IBaseService;
-import domain.entity.unidade.federativa.UnidadeFederativa;
-
-public interface IUnidadeFederativaService extends IBaseService<UnidadeFederativa> {
-
-    UnidadeFederativa ativar(String id, Boolean ativo) throws DomainException;
-}
